updated amenity controller with user role authentication

+++ src/controllers/amenity.controller.ts
@@ -7,6 +7,7 @@ import {
   deleteFileFromCloudinary,
   uploadFileToCloudinary,
 } from "../utils/cloudinary.utils";
+import { Role } from "../types/enum.types";
 const folder = "/amenities";
 
 export const getAll = catchAsync(async (req: Request, res: Response) => {
@@ -58,16 +59,25 @@ export const create = catchAsync(async (req: Request, res: Response) => {
 
 export const update = catchAsync(async (req: Request, res: Response) => {
   const id = req.params.id;
-  const { name, description, user } = req.body;
+  const { name, description } = req.body;
   const file = req.file;
+  const user = req.user;
 
-  const amenity = await Amenity.findOne({ _id: id, user });
+  const amenity = await Amenity.findOne({ _id: id }).populate("user");
+
+  if (!amenity) throw new AppError("Amenity not found", 404, "NOT FOUND");
+
+  //only admin and owner can update
+  if (user.role !== Role.ADMIN || amenity.user._id !== user._id) {
+    throw new AppError("Only admin or owner can update this resource", 400);
+  }
 
   if (!amenity) {
     throw new AppError("Amenity not found", 404);
   }
   if (name) amenity.name = name;
   if (description) amenity.description = description;
+
   if (file) {
     const { path, public_id } = await uploadFileToCloudinary(file, folder);
     await deleteFileFromCloudinary(amenity.icon.public_id);
@@ -86,14 +96,16 @@ export const update = catchAsync(async (req: Request, res: Response) => {
 
 export const remove = catchAsync(async (req: Request, res: Response) => {
   const id = req.params.id;
-  const { user } = req.body;
+  const user = req.user;
   const amenity = await Amenity.findOne({ _id: id }).populate("user");
 
   if (!amenity) {
-    throw new AppError("Amenity not found", 404);
+    throw new AppError("Amenity not found", 404, "NOT FOUND");
   }
-  if (amenity.user._id !== user || amenity.user.role !== "ADMIN")
-    throw new AppError("User access denied", 403);
+
+  if (user.role !== Role.ADMIN || amenity.user._id !== user._id) {
+    throw new AppError("Only admin or owner can update this resource", 400);
+  }
 
   await deleteFileFromCloudinary(amenity.icon.public_id);
   await Amenity.deleteOne();
